src/requirements.txt
requests
aws-lambda-powertools
boto3
awscli
PyJWT
pydantic[dotenv]
fastapi
uvicorn
botocore[crt]
rollbar
mangum